# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Starter API",
        "description": "Laravel Starter API Documentation",
        "contact": {
            "email": "rameespu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/create-user": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create User",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::createUser",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Magic Login Url",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Validation"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout User",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Validation"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::profile",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Validation"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Users",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::getUsers",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication Related APIs"
        },
        {
            "name": "User",
            "description": "User Related APIs"
        }
    ]
}